Working state management (ethy++) & getting rid of initial index file
